integrate Decoder and AsyncDecoder with no extra overhead
