+++ benchmark/profile-decode.ts
@@ -1,7 +1,9 @@
-import { encode, decode } from "../src";
+import { encode, decode, decodeAsync, ExtensionCodec } from "../src";
 
 const data = require("./benchmark-from-msgpack-lite-data.json");
 const encoded = encode(data);
+
+console.log("encoded size:", encoded.byteLength);
 
 console.time("decode #1");
 for (let i = 0; i < 10000; i++) {
@@ -9,8 +11,14 @@ for (let i = 0; i < 10000; i++) {
 }
 console.timeEnd("decode #1");
 
-console.time("decode #2");
-for (let i = 0; i < 10000; i++) {
-  decode(encoded);
-}
-console.timeEnd("decode #2");
+(async () => {
+  const buffers = async function*() {
+    yield encoded;
+  };
+
+  console.time("decodeAsync #1");
+  for (let i = 0; i < 10000; i++) {
+    await decodeAsync(buffers());
+  }
+  console.timeEnd("decodeAsync #1");
+})();
