+++ src/decodeAsync.ts
@@ -1,5 +1,5 @@
 import { ExtensionCodecType, ExtensionCodec } from "./ExtensionCodec";
-import { AsyncDecoder } from "./AsyncDecoder";
+import { Decoder } from "./Decoder";
 
 type DecodeAsyncOptions = Partial<
   Readonly<{
@@ -11,6 +11,6 @@ export async function decodeAsync(
   buffers: AsyncIterable<Uint8Array | ArrayLike<number>>,
   options: DecodeAsyncOptions = {},
 ): Promise<unknown> {
-  const asyncDecoder = new AsyncDecoder(buffers, options.extensionCodec || ExtensionCodec.defaultCodec);
-  return await asyncDecoder.decode();
+  const decoder = new Decoder(options.extensionCodec || ExtensionCodec.defaultCodec);
+  return await decoder.decodeAsync(buffers);
 }
