+++ test/decodeAsync.test.ts
@@ -2,9 +2,23 @@ import assert from "assert";
 import { encode, decodeAsync } from "../src";
 
 describe("decodeAsync", () => {
-  it("decodes [fixarray, nil]", async () => {
+
+  function wrapWithNoisyBuffer(byte: number) {
+    return Uint8Array.from([0x01, byte, 0x02]).subarray(1, 2);
+  }
+
+  it("decodes nil", async () => {
     const asyncBuffers = async function*() {
-      yield [0x91]; // fixarray size=1
+      yield wrapWithNoisyBuffer(0xc0); // nil
+    };
+
+    const object = await decodeAsync(asyncBuffers());
+    assert.deepStrictEqual(object, null);
+  });
+
+  it("decodes fixarray [nil]", async () => {
+    const asyncBuffers = async function*() {
+      yield wrapWithNoisyBuffer(0x91); // fixarray size=1
       yield [0xc0]; // nil
     };
 
@@ -33,9 +47,9 @@ describe("decodeAsync", () => {
     assert.deepStrictEqual(object, 0x1234);
   });
 
-  it("decodes string byte-by-byte", async () => {
+  it("decodes fixstr byte-by-byte", async () => {
     const asyncBuffers = async function*() {
-      yield [0xa3]; // fixstr 3
+      yield [0xa3]; // fixstr size=3
       yield [0x66]; // "f"
       yield [0x6f]; // "o"
       yield [0x6f]; // "o"
@@ -44,10 +58,10 @@ describe("decodeAsync", () => {
     assert.deepStrictEqual(object, "foo");
   });
 
-  it("decodes string byte-by-byte", async () => {
+  it("decodes binary byte-by-byte", async () => {
     const asyncBuffers = async function*() {
       yield [0xc4]; // bin 8
-      yield [0x03]; // size 3
+      yield [0x03]; // bin size=3
       yield [0x66]; // "f"
       yield [0x6f]; // "o"
       yield [0x6f]; // "o"
@@ -56,28 +70,43 @@ describe("decodeAsync", () => {
     assert.deepStrictEqual(object, Uint8Array.from([0x66, 0x6f, 0x6f]));
   });
 
-  it("decodes mixed object byte-by-byte", async () => {
-    const object = {
-      nil: null,
-      true: true,
-      false: false,
-      int: -42,
-      uint64: Number.MAX_SAFE_INTEGER,
-      int64: Number.MIN_SAFE_INTEGER,
-      float: Math.PI,
-      string: "Hello, world!",
-      longString: "Hello, world!\n".repeat(100),
-      binary: Uint8Array.from([0xf1, 0xf2, 0xf3]),
-      array: [1000, 2000, 3000],
-      map: { foo: 1, bar: 2, baz: 3 },
-      timestampExt: new Date(),
-    };
 
+  it("decodes binary with noisy buffer", async () => {
     const asyncBuffers = async function*() {
-      for (const byte of encode(object)) {
-        yield [byte];
-      }
+      yield wrapWithNoisyBuffer(0xc5); // bin 16
+      yield [0x00];
+      yield [0x00]; // bin size=0
     };
-    assert.deepStrictEqual(await decodeAsync(asyncBuffers()), object);
+    const object = await decodeAsync(asyncBuffers());
+    assert.deepStrictEqual(object, new Uint8Array(0));
   });
+
+  // it("decodes mixed object byte-by-byte", async () => {
+  //   const object = {
+  //     nil: null,
+  //     true: true,
+  //     false: false,
+  //     int: -42,
+  //     uint64: Number.MAX_SAFE_INTEGER,
+  //     int64: Number.MIN_SAFE_INTEGER,
+  //     float: Math.PI,
+  //     string: "Hello, world!",
+  //     longString: "Hello, world!\n".repeat(100),
+  //     binary: Uint8Array.from([0xf1, 0xf2, 0xf3]),
+  //     array: [1000, 2000, 3000],
+  //     map: { foo: 1, bar: 2, baz: 3 },
+  //     timestampExt: new Date(),
+  //     map0: {},
+  //     array0: [],
+  //     str0: "",
+  //     bin0: Uint8Array.from([]),
+  //   };
+
+  //   const asyncBuffers = async function*() {
+  //     for (const byte of encode(object)) {
+  //       yield [byte];
+  //     }
+  //   };
+  //   assert.deepStrictEqual(await decodeAsync(asyncBuffers()), object);
+  // });
 });
